+++ src/main/java/qna/domain/repository/AnswerRepository.java
@@ -1,11 +1,14 @@
-package qna.domain;
+package qna.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 import java.util.Optional;
+import qna.domain.Answer;
+import qna.domain.Question;
 
 public interface AnswerRepository extends JpaRepository<Answer, Long> {
+
     List<Answer> findByQuestionAndDeletedFalse(Question question);
 
     Optional<Answer> findByIdAndDeletedFalse(Long id);
