+++ src/main/java/qna/domain/repository/DeleteHistoryRepository.java
@@ -1,12 +1,15 @@
-package qna.domain;
+package qna.domain.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+import qna.domain.ContentType;
+import qna.domain.DeleteHistory;
 
 public interface DeleteHistoryRepository extends JpaRepository<DeleteHistory, Long> {
+
     @Query(value = "select * from delete_history where content_type = :#{#type?.name()}", nativeQuery = true)
     List<DeleteHistory> findByContentType(@Param("type") ContentType type);
