+++ src/main/java/qna/domain/repository/QuestionRepository.java
@@ -1,11 +1,13 @@
-package qna.domain;
+package qna.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 import java.util.Optional;
+import qna.domain.Question;
 
 public interface QuestionRepository extends JpaRepository<Question, Long> {
+
     List<Question> findByDeletedFalse();
 
     Optional<Question> findByIdAndDeletedFalse(Long id);
