+++ src/test/java/qna/domain/UserRepositoryTest.java
@@ -11,6 +11,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.dao.DataIntegrityViolationException;
+import qna.domain.repository.UserRepository;
 
 @DataJpaTest
 public class UserRepositoryTest {
