+++ src/test/java/qna/service/QnaServiceTest.java
@@ -10,10 +10,11 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import qna.common.exception.CannotDeleteException;
 import qna.domain.*;
 
-import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
+import qna.domain.repository.AnswerRepository;
+import qna.domain.repository.QuestionRepository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
